Guide Movement Script
There's still a couple of errors that I really *should* go through an fix up, but for now, it should work well enough.

Tell the script how many stops it is making.

Add waypoints into the inspector for each of those stops.

Customize the speed and wait times if desired by using the inspector of waypoints.

Play the guide movements by checking the "Start Movement" box.