Jump Script Implemented
Fully customizable jump script is live - change gravity and jump strength straight in the inspector to match the needs of the level design. I also separated the rune code into it's own script to prevent possible Git conflicts in the future.

A note about the jump script: The ground checker that is a part of it requires that the surface the player is standing on be a part of the "Ground" mask. I changed the planes in all of the scenes to be "Ground", but any future surfaces you want the player to walk on will need the appropriate layer mask as well.